Turn To Angle Change
Here is an example using the error of the difference between the angle we are at and the angle we want to go to.  This error is used to slow the robot down.  It would be great to use this as a basis and then one of the team can