Adding Lights Python
This python script will be set to run on the raspberry pi that has WPILib Raspberry Pi Vision Image installed.  This script will run "team_color" which is a function to show flame style colors until told otherwise.
Every .1 seconds the program will check with the network tables value of "color" to see if there is a change.  The color value in SmartDashboard will be strings like red, green, blue, off